Refactor NullabilityVisitor/Context out of NotPredicate
