+++ server/src/main/java/io/crate/expression/predicate/NotPredicate.java
@@ -21,9 +21,6 @@
 
 package io.crate.expression.predicate;
 
-import java.util.HashSet;
-import java.util.Set;
-
 import org.apache.lucene.search.BooleanClause;
 import org.apache.lucene.search.BooleanClause.Occur;
 import org.apache.lucene.search.BooleanQuery;
@@ -31,13 +28,10 @@ import org.apache.lucene.search.Query;
 import org.elasticsearch.common.lucene.search.Queries;
 
 import io.crate.data.Input;
-import io.crate.expression.scalar.Ignore3vlFunction;
-import io.crate.expression.scalar.cast.ExplicitCastFunction;
-import io.crate.expression.scalar.cast.ImplicitCastFunction;
+import io.crate.expression.eval.NullabilityVisitor;
 import io.crate.expression.symbol.Function;
 import io.crate.expression.symbol.Literal;
 import io.crate.expression.symbol.Symbol;
-import io.crate.expression.symbol.SymbolVisitor;
 import io.crate.lucene.LuceneQueryBuilder;
 import io.crate.metadata.NodeContext;
 import io.crate.metadata.Reference;
@@ -94,78 +88,7 @@ public class NotPredicate extends Scalar<Boolean, Boolean> {
         return value != null ? !value : null;
     }
 
-
     private final NullabilityVisitor INNER_VISITOR = new NullabilityVisitor();
-
-    private static class NullabilityContext {
-        private final HashSet<Reference> nullableReferences = new HashSet<>();
-        private boolean isNullable = true;
-        private boolean enforceThreeValuedLogic = false;
-
-        void collectNullableReferences(Reference symbol) {
-            nullableReferences.add(symbol);
-        }
-
-        Set<Reference> nullableReferences() {
-            return nullableReferences;
-        }
-
-        boolean enforceThreeValuedLogic() {
-            return enforceThreeValuedLogic;
-        }
-    }
-
-    private static class NullabilityVisitor extends SymbolVisitor<NullabilityContext, Void> {
-
-        private final Set<String> CAST_FUNCTIONS = Set.of(ImplicitCastFunction.NAME, ExplicitCastFunction.NAME);
-
-        @Override
-        public Void visitReference(Reference symbol, NullabilityContext context) {
-            // if ref is nullable and all its parents are nullable
-            if (symbol.isNullable() && context.isNullable) {
-                context.collectNullableReferences(symbol);
-            }
-            return null;
-        }
-
-        @Override
-        public Void visitFunction(Function function, NullabilityContext context) {
-            String functionName = function.name();
-            if (CAST_FUNCTIONS.contains(functionName)) {
-                // Cast functions should be ignored except for the case where the incoming
-                // datatype is an object. There we need to exclude null values to not match
-                // empty objects on the query
-                var a = function.arguments().get(0);
-                var b = function.arguments().get(1);
-                if (a instanceof Reference ref && b instanceof Literal<?>) {
-                    if (ref.valueType().id() == DataTypes.UNTYPED_OBJECT.id()) {
-                        return null;
-                    }
-                }
-            } else if (Ignore3vlFunction.NAME.equals(functionName)) {
-                context.isNullable = false;
-                return null;
-            } else {
-                var signature = function.signature();
-                if (signature.hasFeature(Feature.NON_NULLABLE)) {
-                    context.isNullable = false;
-                } else if (!signature.hasFeature(Feature.NULLABLE)) {
-                    // default case
-                    context.enforceThreeValuedLogic = true;
-                    return null;
-                }
-            }
-            // saves and restores isNullable of the current context
-            // such that any non-nullables observed from the left arg is not transferred to the right arg.
-            boolean isNullable = context.isNullable;
-            for (Symbol arg : function.arguments()) {
-                arg.accept(this, context);
-                context.isNullable = isNullable;
-            }
-            return null;
-        }
-    }
-
 
     @Override
     public Query toQuery(Function input, LuceneQueryBuilder.Context context) {
@@ -194,7 +117,7 @@ public class NotPredicate extends Scalar<Boolean, Boolean> {
         Query innerQuery = arg.accept(context.visitor(), context);
         Query notX = Queries.not(innerQuery);
 
-        NullabilityContext ctx = new NullabilityContext();
+        NullabilityVisitor.NullabilityContext ctx = new NullabilityVisitor.NullabilityContext();
         arg.accept(INNER_VISITOR, ctx);
 
         if (ctx.enforceThreeValuedLogic()) {
